+++ Packages/best.crux.rotation-sensor/CHANGELOG.md
@@ -7,4 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-The entire thing
+### Added
+
+- RotationSensorDefinition, which can be used either:
+  - As a standalone source of rotation rate
+  - In a ProceduralController